refactor(utils): move GetEnvByKey function to os_env.go
- Move GetEnvByKey function from common.go to os_env.go
- Update related test cases from common_test.go to os_env_test.go
